Solution to AoC 2023, day 21, part 2.
Finally! It took a while to get it through my head that I needed to to compute the third derivative (which has a cyclical variation), and use that to compute the values. I just computed every position, which is slow. I'm sure my solution could be easily optimized to compute a whole cycle at once.